# apps/backend/src/main/java/com/candidateupskill/backend/identity/api/LoginRequest.java
package com.candidateupskill.backend.identity.api;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;

public record LoginRequest(
	@NotBlank @Email @Size(max = 320) String email,
	@NotBlank @Size(max = 128) String password
) {
}