# apps/backend/src/main/java/com/candidateupskill/backend/identity/api/SignupRequest.java
package com.candidateupskill.backend.identity.api;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;

public record SignupRequest(
	@NotBlank @Size(max = 160) String name,
	@NotBlank @Email @Size(max = 320) String email,
	@NotBlank @Size(min = 8, max = 128) String password
) {
}